+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -19,6 +19,7 @@
 #include "GuardInForCheck.h"
 #include "GuardInWhileCheck.h"
 #include "HashBufLenCheck.h"
+#include "KeyletBufLenCheck.h"
 #include "RaddrConvBufLenCheck.h"
 #include "RaddrConvPureCheck.h"
 #include "ReserveLimitCheck.h"
@@ -51,6 +52,8 @@ public:
         "hooks-guard-in-while");
     CheckFactories.registerCheck<HashBufLenCheck>(
         "hooks-hash-buf-len");
+    CheckFactories.registerCheck<KeyletBufLenCheck>(
+        "hooks-keylet-buf-len");
     CheckFactories.registerCheck<RaddrConvBufLenCheck>(
         "hooks-raddr-conv-buf-len");
     CheckFactories.registerCheck<RaddrConvPureCheck>(