+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -149,6 +149,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-keylet-buf-len
+  <clang-tidy/checks/hooks-keylet-buf-len>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-raddr-conv-buf-len
   <clang-tidy/checks/hooks-raddr-conv-buf-len>` check.